requirements.txt
docker
fastapi==0.100.0
numpy==1.23.5
uvicorn[standard]
python-multipart
setuptools==59.5.0
wheel
pytorch-lightning==1.2.9
albumentations==0.5.2
hydra-core
webdataset
kornia==0.5.0
pyyaml
tqdm
easydict==1.9.0
scikit-image
scikit-learn
opencv-python
tensorflow
joblib
matplotlib
pandas
tabulate
packaging
wldhx.yadisk-direct
boto3
urllib3
omegaconf
requests